Directory pprint Test
File Mapping: The script then maps files from the ‘Downloads’ directory to their respective file extensions. It does this by listing all files in the ‘Downloads’ directory, splitting each filename to get the file extension, and then appending the filename to the corresponding list in a dictionary (files_mapping). The dictionary uses the file extension as the key and a list of filenames with that extension as the value.

Printing the Result: Finally, it pretty-prints the files_mapping dictionary. This gives you a clear, readable output of which files in the ‘Downloads’ directory have which extensions.